requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
langchain-chroma
langchain-huggingface
sentence-transformers
pypdf
fastapi
uvicorn
python-dotenv
streamlit
